Fix StringView value when running devtool message on JS thread (#122)
After testing devtools profiler in a large app, we noticed that the profiler stop command would crash with an invalid JSON message error. Upon inspection, what happens is the value of `messageView` when inside the `runOnQueue` closure is no longer the right value, which then fails to parse as JSON in the v8 code.

I think StringView is associated to a v8 scope, which would cause it to be garbage collected at the end of the function. Passing the std::string to the closure instead fixes the problem.